+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - `@rarimo/rarime`:
-  - add `issuerDID` to filter credentials by issuer
+  - add `issuerDid` to filter credentials by issuer
 
 
 ## [0.6.0] - 2023-10-19